Migrate general casgo tests, separate integration tests
